refactor(tests): migrated remaining tests to ava - updated package to remove mocha dep - renamed empty tests so it does not get picked up
