+++ src/utils/utils.spec.js
@@ -1,25 +1,19 @@
-
+import test from 'ava'
 import { isEmpty, getNameAndExtension } from './utils'
 
-describe('utils', function () {
-  describe('isEmpty', () => {
-    it('should determine if a string is empty', function () {
-      const emptyInput = ''
-      const notEmptyInput = 'foo'
-      assert.strictEqual(isEmpty(emptyInput), true)
-      assert.strictEqual(isEmpty(notEmptyInput), false)
-    })
+test('utils: isEmpty: should determine if a string is empty', t => {
+  const emptyInput = ''
+  const notEmptyInput = 'foo'
+  t.deepEqual(isEmpty(emptyInput), true)
+  t.deepEqual(isEmpty(notEmptyInput), false)
+})
 
-    it('should not fail with a non string input ', function () {
-      const input = {foo: 42}
-      assert.strictEqual(isEmpty(input), false)
-    })
-  })
+test('utils: isEmpty: should not fail with a non string input ', t => {
+  const input = {foo: 42}
+  t.deepEqual(isEmpty(input), false)
+})
 
-  describe('getNameAndExtension', () => {
-    it('should return the name and the extension from a dotted string', function () {
-      const input = 'foo.bar.STL'
-      t.deepEqual(getNameAndExtension(input), {name: 'foo.bar.STL', ext: 'stl'})
-    })
-  })
+test('utils: getNameAndExtension: should return the name and the extension from a dotted string', t => {
+  const input = 'foo.bar.STL'
+  t.deepEqual(getNameAndExtension(input), {name: 'foo.bar.STL', ext: 'stl'})
 })
